backend-apis/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
greenlet # Required by SQLAlchemy for async ORM features, good to have if future plans involve asyncpg
python-multipart # Required for form data handling

# Testing dependencies
pytest
pytest-asyncio
httpx
pytest-cov
